fixup build
